guide/requirements.txt
selenium==4.15.2
requests==2.31.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
fuzzywuzzy==0.18.0
python-dotenv==1.0.0
azure-storage-blob==12.19.0
